refactor(infra): move gpu-effects from lib/ to infrastructure/
WebGPU effect pipeline is platform-coupled code, so it belongs in the
infrastructure layer rather than lib/. Removes the re-export facade
that previously existed only to bridge the boundary rule.

First slice of a planned migration that pulls the remaining gpu-* /
analysis / audio modules out of lib/ so the layering reads as:
lib/ = generic utilities, infrastructure/ = platform adapters.
