+++ src/features/effects/components/panels/gpu-effect-panel.tsx
@@ -10,7 +10,7 @@ import {
   SelectValue,
 } from '@/components/ui/select'
 import type { ItemEffect, GpuEffect } from '@/types/effects'
-import type { GpuEffectDefinition } from '@/infrastructure/gpu/effects'
+import type { GpuEffectDefinition } from '@/infrastructure/gpu-effects'
 import { KeyframeToggle } from '@/features/effects/deps/keyframes-contract'
 import type { AnimatableProperty } from '@/types/keyframe'
 import { ColorPicker, PropertyRow, SliderInput } from '@/shared/ui/property-controls'
