+++ src/features/effects/components/panels/gpu-wheels-panel.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { Eye, EyeOff, Trash2, RotateCcw } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import type { ItemEffect, GpuEffect } from '@/types/effects'
-import type { GpuEffectDefinition } from '@/infrastructure/gpu/effects'
+import type { GpuEffectDefinition } from '@/infrastructure/gpu-effects'
 import { KeyframeToggle } from '@/features/effects/deps/keyframes-contract'
 import type { AnimatableProperty } from '@/types/keyframe'
 import { PropertyRow, SliderInput } from '@/shared/ui/property-controls'
