+++ src/features/export/utils/canvas-effects.ts
@@ -8,7 +8,7 @@ import type { ItemKeyframes } from '@/types/keyframe'
 import type { ItemEffect, GpuEffect } from '@/types/effects'
 import type { AdjustmentItem, TimelineItem } from '@/types/timeline'
 import { createLogger } from '@/shared/logging/logger'
-import type { EffectsPipeline, GpuEffectInstance } from '@/infrastructure/gpu/effects'
+import type { EffectsPipeline, GpuEffectInstance } from '@/infrastructure/gpu-effects'
 import { applyMasks, type MaskCanvasSettings } from './canvas-masks'
 import type { CanvasPool } from './canvas-pool'
 import { resolveAnimatedColorEffects } from '@/features/export/deps/keyframes'
