+++ src/features/preview/hooks/use-preview-renderer-controller.ts
@@ -1102,7 +1102,7 @@ export function usePreviewRendererController({
     if (!FAST_SCRUB_RENDERER_ENABLED) return
     void (async () => {
       try {
-        const { EffectsPipeline } = await import('@/infrastructure/gpu/effects')
+        const { EffectsPipeline } = await import('@/infrastructure/gpu-effects')
         const device = await EffectsPipeline.requestCachedDevice()
         if (device) {
           const warmPipeline = await EffectsPipeline.create()
