+++ src/infrastructure/gpu-effects/common.ts
