+++ src/infrastructure/gpu-effects/effects-pipeline.test.ts
