+++ src/infrastructure/gpu-effects/effects-pipeline.ts
