+++ src/infrastructure/gpu-effects/effects/stylize.ts
