+++ src/infrastructure/gpu-effects/index.test.ts
