+++ src/infrastructure/gpu-effects/types.ts
