+++ src/infrastructure/thumbnails/gpu-thumbnail-renderer.ts
@@ -8,7 +8,7 @@
  * Lazy singleton — created on first use.
  */
 
-import { EffectsPipeline } from '@/infrastructure/gpu/effects'
+import { EffectsPipeline } from '@/infrastructure/gpu-effects'
 import { CompositorPipeline } from '@/infrastructure/gpu/compositor'
 import { MaskTextureManager } from '@/infrastructure/gpu/masks'
